Update docker-compose, Dockerfile, and other changes for xrx-reasoning service
