+++ simple-agent/reasoning/Dockerfile
@@ -1,10 +1,11 @@
 FROM python:3.10-slim
 
-WORKDIR /agent
+WORKDIR /app
 
-COPY requirements.txt requirements.txt
+COPY reasoning/requirements.txt requirements.txt
 RUN pip install -r requirements.txt
 
-COPY ./app ./
+COPY reasoning/app .
+COPY xrx-core/agent_framework /app/agent_framework
 
 CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8003"]
